Support hourly logrotate directory on EL9 again.
When EL9 switched to using systemd (quite correctly ) in #216 support
for a special directory for hourly logrotations was removed.

Add that functionality back with a new unit and timer
* `logrotate-hourly.service`
* `logrotate-hourly.timer`

While this could be achieved by reducing the period of existing timer
to hourly I presume the motivation for this was to avoid processing
all logrotate directives every hour.