Allow disabling logrotate-hourly when using systemd
Similar to the option to disable hourly logrotation when using Cron
(`logrotate::ensure_cron_hourly => absent`) this change allows to
optionally disable hourly logrotation when using systemd. Previously
this was not possible; when the systemd timer/service for logrotate is
managed then the hourly logrotation timer/service is also present and
enabled.

With this change we introduce the class parameter
`logrotate::ensure_systemd_timer_hourly` which defaults to "present",
but can be set to "absent" so that any hourly logrotation timer,
service and drop-in are removed.

This can be useful in cases where the usage of "flock" conflicts with
"postrotate" commands that inadvertently take ownership of the lock file
and keep it open indefinitely; this results in logrotate blocking
completely and eventually failing when the lock expires after six hours.